Invert location of deframer fuzzing harness
Put it inside the crate and build it unconditionally so
breakage is visible without needing to run `cargo fuzz build`.

Eliminate the exports under `internals` that exist solely
for it.
