+++ fuzz/fuzzers/deframer.rs
@@ -3,39 +3,6 @@
 extern crate libfuzzer_sys;
 extern crate rustls;
 
-use std::io;
+use rustls::internal::fuzzing::fuzz_deframer;
 
-use rustls::internal::msgs::deframer;
-use rustls::internal::msgs::message::Message;
-use rustls::internal::record_layer::RecordLayer;
-
-fuzz_target!(|data: &[u8]| {
-    let mut buf = deframer::DeframerVecBuffer::default();
-    let mut dfm = deframer::MessageDeframer::default();
-    if dfm
-        .read(&mut io::Cursor::new(data), &mut buf)
-        .is_err()
-    {
-        return;
-    }
-    buf.has_pending();
-
-    let mut rl = RecordLayer::new();
-    let mut discard = 0;
-
-    loop {
-        let mut borrowed_buf = buf.borrow();
-        borrowed_buf.queue_discard(discard);
-
-        let res = dfm.pop(&mut rl, None, &mut borrowed_buf);
-        discard = borrowed_buf.pending_discard();
-
-        if let Ok(Some(decrypted)) = res {
-            Message::try_from(decrypted.message).ok();
-        } else {
-            break;
-        }
-    }
-
-    buf.discard(discard);
-});
+fuzz_target!(|bytes: &[u8]| { fuzz_deframer(bytes) });
